Address review feedback.
Move tests from `resource_internal_test.go` to `components_test.go`.
